chore: split message types into own files
