+++ src/types/messages/SetToken.ts
@@ -1,0 +1,16 @@
+import { MessageType } from '../MessageType'
+import { Message } from './Message'
+
+/**
+ * Set the token in the iframe
+ */
+export interface SetTokenMessage extends Message {
+  /**
+   * @inheritdoc
+   */
+  type: MessageType.SET_TOKEN
+  /**
+   * The token to set
+   */
+  payload: string
+}
