+++ src/types/messages/index.ts
@@ -1,0 +1,18 @@
+import { OfcpConfig } from '../Config'
+
+export * from '../MessageType'
+export * from './Message'
+export * from './RequestToken'
+export * from './SetToken'
+export * from './StylesRequest'
+export * from './StylesSendData'
+export * from './CallParentEvent'
+export * from './MetaRequest'
+export * from './MetaSendData'
+export * from './Ping'
+export * from './Pong'
+export * from './CloseFullScreen'
+export * from './OpenFullScreen'
+export * from './DestroyFullScreen'
+export * from './FullScreenOpened'
+export * from './FullScreenClosed'
